Switch from lncli to the gRPC API
Switched to gRPC to interface with lnd.
All changes from lncli to the gRPC API are straightforward and mostly consist in dealing with protos rather than JSON-formatted data. One exception is the generate_invoice function, especially where the formatting of r_hash is confusing to be later used in LookupInvoice. The workaround implemented is to take the payment_request and then DecodePayReq which gives a payment_hash.