Limit fees based on lnd defaults
For each hop where fees have to be paid (which excludes the first hop)
we compute the fees based on the default settings of lnd. The actual
fees are compared to this. The route is rejected if the actual fees are
higher than the default lnd fees multiplied by a configurable factor
(default: 10).

For a route with n hops the upper limit is:
factor * (n-1) * (1000 + (amount * 1 / 1000000))

Note: The amount considered to compute the limit is the amount sent for
the rebalance. In reality fees of previous hops also contribute to the
fee computation of successor hops.

Fixes #9