use final amount in QueryRoutes
This way lnd can make use of its mission control system
so that we only get routes that are not banned for the
given amount. Without this fix we might ask for an
amount where the fee we have to pay for the last hop back
to our own node (which is not part of the QueryRoutes
interaction!) is not regarded.

For the existing route extension logic to work, we need
to use the original amounts in the route, though. This
is done by subtracting the added fee again.

fixes #67