add -A / --adjust-amount-to-limits
If you rebalance with -A set, the amount is automatically reduced based on the --min-local
and --min-remote limits. As an example, if you call with `-t 111 -a 50000` where the target
channel 111 already has at least --min-local sats on your side of the channel, the script
will reduce the amount to 0 and stop. If the target channel only lacks 5,000sat to reach
--min-local, the amount will be reduced to 5,000sat. Depending on --min-amount (defaulting
to 10,000sat), the script may stop because the amount is too small. If the target channel
needs 50,000sat or more to reach --min-local, 50,000sat is used.

fixes #215